requirements_lint.txt
# pip requirements
yamllint
pylint
ansible-core
ansible-lint
requests
zeep
